chore(valid-title): adjust default for `disallowedWords` option
`@typescript-eslint` v3 changed the structure of their types, meaning
that we need to type the options for `valid-title`.

This in turn means we have to provide a direct default value for the
`disallowedWords` option, in turn meaning we have to adjust our test to
ensure proper coverage instead of relying on the `default` property.