fix: make no-jest-import error message more accurate
It isn't quite accurate to say that Jest doesn't export anything. It exports the `run` and `runCLI` functions, even though it isn't generally recommended to use them.

This commit updates the message to match the language in https://jestjs.io/docs/jest-object ("The `jest` object is automatically in scope within every test file").

I'll follow this commit with an update to the no-jest-import.md file too.